backend/requirements.txt
# Core
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.5
starlette==0.46.2

# CORS, HTTP, and Utilities
httpx==0.28.1
requests==2.32.3
anyio==4.9.0
sniffio==1.3.1
click==8.2.1

# HTML Parsing and Scraping
beautifulsoup4==4.13.4
soupsieve==2.7

# Transformers & NLP
transformers==4.52.3
tokenizers==0.21.1
torch==2.7.0
sentence-transformers==4.1.0
safetensors==0.5.3
huggingface-hub==0.32.2
filelock==3.18.0

# FAISS for semantic search
faiss-cpu==1.11.0
numpy==2.2.6
